# api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "lct api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "BSD 3-Clause License",
            "url": "https://raw.githubusercontent.com/EgorTarasov/true-tech/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "api.lct.larek.tech",
    "basePath": "/ //no-lint.",
    "paths": {
...
        "/data/predict": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "получение предсказание аварийных ситуаций и информации об объекте",
                "parameters": [
                    {
                        "description": "predictionRequest",
                        "name": "predictionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.predictionRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/issue/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "создание инцидента",
                "parameters": [
                    {
                        "description": "incidentCreateRequest",
                        "name": "incidentCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.incidentCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/issue/heating-point/{unom}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "получение информации об инцидетнах по уникальному номеру объекта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "unom",
                        "name": "unom",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Incident"
                            }
                        }
                    }
                }
            }
        },
        "/issue/id/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "получение информации об инцидетне по его идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Incident"
                        }
                    }
                }
            }
        },
        "/issue/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "получение информации об инцидетнах",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Incident"
                            }
                        }
                    }
                }
            }
        },
        "/reports/create": {
            "post": {
                "description": "Создание отчета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Создание отчета",
                "parameters": [
                    {
                        "description": "reportCreate",
                        "name": "reportCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reports.CreateReport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/reports/my": {
            "get": {
                "description": "Получение списка отчетов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Получение списка отчетов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reports.Report"
                            }
                        }
                    }
                }
            }
        },
        "/reports/{id}": {
            "get": {
                "description": "Получение отчета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Получение отчета",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reports.Report"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.incidentCreateRequest": {
            "type": "object",
            "required": [
                "priority",
                "status",
                "title"
            ],
            "properties": {
                "priority": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.predictionRequest": {
            "type": "object",
            "required": [
                "admArea",
                "endDate",
                "startDate",
                "threshold"
            ],
            "properties": {
                "admArea": {
                    "type": "string"
                },
                "endDate": {
                    "description": "string in format dd-mm-yyyy",
                    "type": "string"
                },
                "startDate": {
                    "description": "string in format dd-mm-yyyy",
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
        "models.Address": {
            "type": "object",
            "required": [
                "municipalDistrict"
            ],
            "properties": {
                "address": {
                    "description": "Полный Адрес в реестре.",
                    "type": "string"
                },
                "border": {
                    "description": "Граница объекта на карте."
                },
                "center": {
                    "description": "Центр объекта на карте.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Point"
                        }
                    ]
                },
                "municipalDistrict": {
                    "description": "Район округ.",
                    "type": "string"
                },
                "unom": {
                    "description": "Уникальный номер объекта недвижимости.",
                    "type": "integer"
                }
            }
        },
...
        "models.DispatchServices": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "centerAddress": {
                    "type": "string"
                },
                "consumer": {
                    "type": "string"
                },
                "consumerFullAddress": {
                    "type": "string"
                },
                "consumerGroup": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "dispatchNumber": {
                    "type": "string"
                },
                "externalID": {
                    "type": "integer"
                },
                "heatDispatcherNumber": {
                    "type": "string"
                },
                "unom": {
                    "type": "integer"
                },
                "События": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "external_closed_at": {
                    "type": "string"
                },
                "external_created_at": {
                    "type": "string"
                },
                "external_ended_at": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "unom": {
                    "type": "integer"
                }
            }
        },
...
        "models.Graph": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GraphDataPoint"
                    }
                }
            }
        },
        "models.GraphDataPoint": {
            "type": "object",
            "properties": {
                "temp": {
                    "type": "number"
                },
                "time": {
                    "description": "время (1 час, 3 часа и т.п.)",
                    "type": "integer"
                }
            }
        },
        "models.HeatingPoint": {
            "type": "object",
            "properties": {
                "balance_holder": {
                    "type": "string"
                },
                "commissioning_date": {},
                "consumer_full_address": {
                    "$ref": "#/definitions/models.Address"
                },
                "consumers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Address"
                    }
                },
                "heating_point_full_address": {
                    "$ref": "#/definitions/models.Address"
                },
                "heating_point_location_type": {
                    "type": "string"
                },
                "heating_point_number": {
                    "type": "string"
                },
                "heating_point_src": {
                    "type": "string"
                },
                "heating_point_type": {
                    "type": "string"
                },
                "mkd_consumers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MKDConsumer"
                    }
                },
                "municipal_district": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "state_consumers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StateConsumer"
                    }
                },
                "unom": {
                    "type": "integer"
                }
            }
        },
...
        "models.Incident": {
            "type": "object",
            "properties": {
                "closedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dispatchServices": {
                    "$ref": "#/definitions/models.DispatchServices"
                },
                "graph": {
                    "$ref": "#/definitions/models.Graph"
                },
                "heatingPoint": {
                    "$ref": "#/definitions/models.HeatingPoint"
                },
                "id": {
                    "type": "integer"
                },
                "measurements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Measurement"
                    }
                },
                "mkdConsumer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MKDConsumer"
                    }
                },
                "openedAt": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "relatedObjects": {
                    "$ref": "#/definitions/models.UnomResult"
                },
                "stateConsumers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StateConsumer"
                    }
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "unom": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.MKDConsumer": {
            "type": "object",
            "properties": {
                "PassengerElevator": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "apartments": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "deprecation": {
                    "type": "number"
                },
                "district": {
                    "type": "string"
                },
                "entrances": {
                    "type": "integer"
                },
                "externalID": {
                    "type": "integer"
                },
                "feature": {
                    "type": "integer"
                },
                "floors": {
                    "type": "integer"
                },
                "fullAddress": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "projectSeries": {},
                "roofCleaningSequence": {
                    "type": "integer"
                },
                "roofMaterials": {
                    "type": "integer"
                },
                "serviceElevator": {
                    "type": "integer"
                },
                "state": {
                    "type": "integer"
                },
                "totalArea": {
                    "type": "number"
                },
                "totalLivingArea": {
                    "type": "number"
                },
                "totalNonLivingArea": {},
                "typesOfHousingStock": {
                    "type": "integer"
                },
                "unom": {
                    "type": "integer"
                },
                "wallMaterial": {
                    "type": "integer"
                },
                "События": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                }
            }
        },
        "models.Measurement": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "consumer": {
                    "description": "Бюджет / МКД",
                    "type": "string"
                },
                "coolantInput": {
                    "type": "number"
                },
                "coolantOutput": {
                    "type": "number"
                },
                "coolantSuppliedDifferenceLeakage": {
                    "type": "number"
                },
                "datetime": {
                    "type": "string"
                },
                "energyConsumption": {
                    "type": "number"
                },
                "heatingPoint": {
                    "type": "string"
                },
                "measurementDate": {
                    "type": "string"
                },
                "measuringDeviceBrand": {
                    "type": "string"
                },
                "measuringDeviceHours": {
                    "type": "number"
                },
                "unom": {
                    "type": "integer"
                }
            }
        },
...
        "models.Point": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.StateConsumer": {
            "type": "object",
            "properties": {
                "admDistrict": {
                    "type": "string"
                },
                "area": {},
                "floors": {},
                "material": {
                    "type": "string"
                },
                "municupalDistrict": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "propertyClass": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "unom": {
                    "type": "integer"
                },
                "События": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                }
            }
        },
...
        "models.UnomResult": {
            "type": "object",
            "properties": {
                "consumer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MKDConsumer"
                    }
                },
                "dispatchService": {
                    "$ref": "#/definitions/models.DispatchServices"
                },
                "heatingPoint": {
                    "$ref": "#/definitions/models.HeatingPoint"
                },
                "unom": {
                    "type": "integer"
                }
            }
        },
...
        "reports.CreateReport": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "reports.Report": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "s3Key": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/reports.ReportState"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "reports.ReportState": {
            "type": "string",
            "enum": [
                "in_progress",
                "ready"
            ],
            "x-enum-varnames": [
                "ReportStateInProgress",
                "ReportStateReady"
            ]
        },
...
    }
}